+++ src/app/[locale]/journal/page.tsx
